requirement.txt
dnspython
requests
dotenv
